(refactor) Merged routes, controller files
