+++ server/src/controllers/clientController.js
@@ -16,7 +16,8 @@ export default {
 
     db.connection.query(query, data, function(error, results, fields) {
       if (error) {
-        console.log('error: ', error, 'faild to insert new user');
+        console.log('error code: ' + error.code +
+                    ', faild to insert new user');
       } else {
         console.log('(signUp)new customer successfully registered');
         res.send(results);
@@ -50,27 +51,33 @@ export default {
   },
 
   getAllContracts: function(req, res) {
-    var client_Email = req.body.client_Email;
+
+    var client_Email = req.params.client_Email;
     var client_Index;
 
-    var firstQ = "SELECT client_Index FROM Client where client_Email=" + client_Email;
-    db.connection.query(firstQ, function(error, results, fields) {
+    console.log('client_Email: ' + client_Email);
+
+    var firstQ = "SELECT client_Index FROM Client where client_Email=?";
+    var firstData = [client_Email];
+
+    db.connection.query(firstQ, firstData, function(error, results, fields) {
       if (error) {
         console.log('error code: ' + error.code +
                     ', failed to retreive client_Index from Client');
       } else {
-        console.log('query: ' + query);
+
+        console.log('query: ' + firstQ);
         console.log(results);
 
         client_Index = results[0];
-
-        var secondQ = "SELECT * FROM deal WHERE client_Index=" + client_Index;
-        db.connection.query(secondQ, function(error, results, fields) {
+        var secondQ = "SELECT * FROM deal WHERE client_Index=?";
+        var secondData = [client_Index];
+        db.connection.query(secondQ, secondData, function(error, results, fields) {
           if (error) {
             console.log('error code: ' + error.code +
                         ', failed to retreive all deals');
           } else {
-            console.log('query: ' + query);
+            console.log('query: ' + secondQ);
             console.log(results);
             res.send(results);
           }
@@ -80,30 +87,33 @@ export default {
   },
 
   getContract: function(req, res) {
-    var client_Email = req.body.client_Email;
-    var booking_Num = req.body.booking_Num;
+
+    var client_Email = req.params.client_Email;
+    var booking_Num = req.params.booking_Num;
     var client_Index;
 
-    var firstQ = "SELECT client_Index FROM Client where client_Email=" + client_Email;
-    db.connection.query(firstQ, function(error, results, fields) {
+    var firstQ = "SELECT client_Index FROM Client where client_Email=?";
+    var firstData = [client_Email];
+
+    db.connection.query(firstQ, firstData, function(error, results, fields) {
       if (error) {
         console.log('error code: ' + error.code +
                     ', failed to retreive client_Index from Client');
       } else {
-        console.log('query: ' + query);
+        console.log('query: ' + firstQ);
         console.log(results);
 
         client_Index = results[0];
 
         // to avoid async problem
-        var secondQ = "SELECT * FROM deal WHERE client_Index=" + client_Index +
-                      "AND booking_Num=" + booking_Num;
-        db.connection.query(secondQ, function(error, results, fields) {
+        var secondQ = "SELECT * FROM deal WHERE client_Index=? AND booking_Num=?";
+        var secondData = [client_Index, booking_Num];
+        db.connection.query(secondQ, secondData, function(error, results, fields) {
           if (error) {
             console.log('error code: ' + error.code +
                         ', failed to retreive all deals');
           } else {
-            console.log('query: ' + query);
+            console.log('query: ' + secondQ);
             console.log(results);
             res.send(results);
           }
@@ -113,7 +123,8 @@ export default {
   },
 
   makeContract: function(req, res) {
-    var client_Email = req.body.client_Email;
+
+    var client_Email = req.params.client_Email;
     var client_Index;
 
     var checkIn_Date = req.body.checkIn_Date;
@@ -122,31 +133,34 @@ export default {
     var subArea_Name = req.body.subArea_Name;
 
     var bid_Price = req.body.bid_Price;
-    var bid_StartTime = req.body.bid_StartTime;
-    var bid_EndTime = req.body.bid_EndTime;
+    // var bid_StartTime = req.body.bid_StartTime;
+    // var bid_EndTime = req.body.bid_EndTime;
     var bid_Transaction = 1;
 
-    var firstQ = 'SELECT client_Index from Client WHERE client_Email=' + client_Email;
-    db.connection.query(firstQ, function(error, results, fields) {
+    var firstQ = 'SELECT client_Index from Client WHERE client_Email=?';
+    var firstData = [client_Email];
+
+    db.connection.query(firstQ, firstData, function(error, results, fields) {
       if (error) {
         console.log('error code: ' + error.code);
       } else {
-        console.log('query: ' + query);
-        client_Index = results[0];
-        console.log('query result: ' + client_Index);
-
+        console.log('query: ' + firstQ, firstData);
+        client_Index = results[0].client_Index;
+        console.log('client_Index: ', client_Index);
 
         var query1 = 'INSERT INTO deal SET ';
         var query2 = 'client_Index=?, checkIn_Date=?, checkOut_Date=?, mainArea_Name=?, subArea_Name=?, ';
-        var query3 = 'bid_Price=?, bid_StartTime=?, bid_EndTime=?, bid_Transaction=?';
-        var query4 = [client_Index, checkIn_Date, checkOut_Date, mainArea_Name, subArea_Name, bid_Price, bid_StartTime, bid_EndTime, bid_Transaction];
+        var query3 = 'bid_Price=?, bid_StartTime=now(), bid_EndTime=now()+interval 1 day, bid_Transaction=?';
+        var secondData = [client_Index, checkIn_Date, checkOut_Date, mainArea_Name, subArea_Name, bid_Price, bid_Transaction];
 
-        var secondQ = query1 + query2 + query3 + query4;
-        db.connection.query(secondQ, function(error, results, fields) {
+        var secondQ = query1 + query2 + query3;
+        console.log(secondQ, secondData);
+
+        db.connection.query(secondQ, secondData, function(error, results, fields) {
           if (error) {
             console.log('error code: ' + error.code);
           } else {
-            console.log('query: ' + query);
+            console.log('query: ' + secondQ, secondData);
             console.log(results);
             // give back the inserted result
             res.send(results);
@@ -166,29 +180,27 @@ export default {
   },
 
   updateInfo: function(req, res) {
+    var client_Email = req.params.client_Email;
+
     var client_ID = req.body.client_ID;
     var client_PW = req.body.client_PW;
     var client_Name = req.body.client_Name;
-    var client_Email = req.body.client_Email;
     var billingInfo = req.body.billingInfo;
     var member = req.body.member;
 
-    var query_header = 'UPDATE Client SET ';
-    var query1 = 'client_PW=' + client_PW + ', ';
-    var query2 = 'client_Name=' + client_Name + ', ';
-    var query3 = 'client_Email=' + client_Email + ', ';
-    var query4 = 'billingInfo=' + billingInfo + ', ';
-    var query5 = 'member=' + member + ' ';
-    var query_trailer = 'WHERE Client_ID=' + Client_ID;
+    var query1 = 'UPDATE Client SET ';
+    var query2 = 'client_PW=?, client_Name=?, client_Email=?, billingInfo=?, member=? WHERE client_ID=?';
+    var data = [client_PW, client_Name, client_Email, billingInfo, member, client_ID];
 
-    var query = query_header + query1 + query2 + query3 + query4 + query5 + query_trailer;
+    var query = query1 + query2;
 
-    db.connection.query(query, function(error, results, fields) {
+    db.connection.query(query, data, function(error, results, fields) {
       if (error) {
         console.log('error code: ' + error.code +
                     ', failed to update client information');
       } else {
-        console.log('query: ' + query);
+
+        console.log('query: ' + query, data);
         console.log(results);
         res.send(results);
       }
