+++ server/src/controllers/hotelController.js
@@ -5,13 +5,16 @@ export default {
     var hotel_ID = req.body.hotel_ID;
     var hotel_PW = req.body.hotel_PW;
     var hotel_Name = req.body.hotel_Name;
-    var hotel_Location = req.body.hotel_Location;
+    var hotel_Address = req.body.hotel_Address;
+    var mainArea_Name = req.body.mainArea_Name;
+    var subArea_Name = req.body.subArea_Name;
     var hotel_Rate = req.body.hotel_Rate;
     var mgr_Name = req.body.mgr_Name;
 
     var query1 = 'INSERT INTO Hotel SET ';
-    var query2 = 'hotel_ID=?, hotel_PW=?, hotel_Name=?, hotel_Location=?, hotel_Rate=?, mgr_Name=?';
-    var data = [hotel_ID, hotel_PW, hotel_Name, hotel_Location, hotel_Rate, mgr_Name];
+    var query2 = 'hotel_ID=?, hotel_PW=?, hotel_Name=?, hotel_Address=?, mainArea_Name=?, subArea_Name=?, hotel_Rate=?, mgr_Name=?';
+    var data = [hotel_ID, hotel_PW, hotel_Name, hotel_Address, mainArea_Name, subArea_Name, hotel_Rate, mgr_Name];
+
     var query = query1 + query2;
 
     db.connection.query(query, data, function(error, results, fields) {
@@ -28,12 +31,15 @@ export default {
 
   signIn: function(req, res) {
     // encryption? later...
-    var id = req.body.Hotel_ID;
-    var pwd = req.body.Hotel_PW;
+    var hotel_ID = req.body.hotel_ID;
+    var hotel_PW = req.body.hotel_PW;
 
-    var query = 'SELECT Hotel_ID, Hotel_PW from Hotel';
+    console.log('hotel_ID: ' + hotel_ID);
 
-    db.connection.query(query, function(error, results, fields) {
+    var query = 'SELECT * from Hotel where hotel_ID=? AND hotel_PW=?';
+    var data = [hotel_ID, hotel_PW];
+
+    db.connection.query(query, data, function(error, results, fields) {
 
       if (error) {
         console.log('error code: ' + error.code +
@@ -118,31 +124,30 @@ export default {
   },
 
 
-  update: function(req, res) {
+  updateInfo: function(req, res) {d
     var hotel_ID = req.params.hotel_ID;
 
     var hotel_PW = req.body.hotel_PW;
     var hotel_Name = req.body.hotel_Name;
-    var hotel_Location = req.body.hotel_Location;
+    var hotel_Address = req.body.hotel_Address;
+    var mainArea_Name = req.body.mainArea_Name;
+    var subArea_Name = req.body.subArea_Name;
     var hotel_Rate = req.body.hotel_Rate;
     var mgr_Name = req.body.mgr_Name;
 
-    var query_header = 'UPDATE Hotel SET ';
-    var query1 = 'hotel_PW=' + hotel_PW + ', ';
-    var query2 = 'hotel_Name=' + hotel_Name + ', ';
-    var query3 = 'hotel_Location=' + hotel_Location + ', ';
-    var query4 = 'hotel_Rate=' + hotel_Rate + ', ';
-    var query5 = 'mgr_Name=' + mgr_Name + ' ';
-    var query_trailer = 'WHERE hotel_ID=' + hotel_ID;
+    var query1 = 'UPDATE Hotel SET ';
+    var query2 = 'hotel_PW=?, hotel_Name=?, hotel_Address=?, mainArea_Name=?, subArea_Name=?, hotel_Rate=?, mgr_Name=? ';
+    var query3 = 'WHERE hotel_ID=?';
+    var data = [hotel_PW, hotel_Name, hotel_Address, mainArea_Name, subArea_Name, hotel_Rate, mgr_Name, hotel_ID];
 
-    var query = query_header + query1 + query2 + query3 + query4 + query5 + query_trailer;
+    var query = query1 + query2 + query3;
 
-    db.connection.query(query, function(error, results, fields) {
+    db.connection.query(query, data, function(error, results, fields) {
       if (error) {
         console.log('error code: ' + error.code +
                     ', failed to update hotel information');
       } else {
-        console.log('query: ' + query);
+        console.log('query: ' + query, data);
         console.log(results);
         res.send(results);
       }
